Display calibration view in Mainscreen (doesn't do anything yet)
