+++ microapps/ArWorldScaleApp/app/src/main/java/com/arcgismaps/toolkit/arworldscaleapp/screens/MainScreen.kt
@@ -20,13 +20,22 @@ package com.arcgismaps.toolkit.arworldscaleapp.screens
 
 import android.util.Log
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import com.arcgismaps.location.SystemLocationDataSource
 import com.arcgismaps.mapping.ArcGISScene
 import com.arcgismaps.mapping.BasemapStyle
 import com.arcgismaps.mapping.Viewpoint
 import com.arcgismaps.toolkit.ar.WorldScaleSceneView
+import com.arcgismaps.toolkit.ar.internal.WorldScaleCalibrationViewDefaults
 
 @Composable
 fun MainScreen() {
@@ -37,11 +46,26 @@ fun MainScreen() {
             scale = 10e7
         )
     }
+    var displayCalibrationView by remember { mutableStateOf(false) }
+
     WorldScaleSceneView(
         arcGISScene = arcGISScene,
         modifier = Modifier.fillMaxSize(),
         onInitializationStatusChanged = {
             Log.d("ArWorldScaleApp", "Initialization status changed: $it")
         }
-    )
+    ) {
+        if (displayCalibrationView) {
+            CalibrationView(
+                onDismiss = { displayCalibrationView = false },
+                modifier = Modifier,
+                colorScheme = WorldScaleCalibrationViewDefaults.colorScheme(),
+                typography = WorldScaleCalibrationViewDefaults.typography()
+            )
+        } else {
+            FloatingActionButton(onClick = {displayCalibrationView = true}) {
+                Icon(Icons.Default.Edit, contentDescription = "Calibration View button")
+            }
+        }
+    }
 }
